Significantly improve annotate_tips performance
The current implementation of annotate_tips calls annotate_node for each
annotation in the metadata structure. However, annotate_node calls
find_node_with_taxon which has a performance complexity that increases
with the size of the tree. For very large, complex trees, the amount
of time spent searching the tree with this function becomes significant.

This commit replaces the implementation of annotate_tips such that it
no longer calls annotate_node to perform a search. Instead, annotate_tips
now iterates through the nodes in the tree with leaf_node_iter and checks
to see if the node's taxon label is in the annotations structure.

Effectively this swaps the loops around; rather than iterating over the
metadata and searching the tree for the node to label, we instead loop over
the tree just once, and check each node to see if there are any annotations
to be applied. This removes the need to search within the tree at all.

This has been tested in Cardiff. We have verfied that the md5sum of
the tree created before and after this commit are exactly the same.
On a recent COG-UK tree, this reduces compute time from 18500 seconds
to just 45 seconds.

Co-authored-by: Matt Bull <bullmj2@gmail.com>